feat: show selection on deep linking launch
